chromium/third_party/blink/web_tests/external/wpt/signed-exchange/subresource/sxg-subresource-header-integrity-mismatch.tentative-expected.txt

This is a testharness.js-based test.
FAIL Subresource signed exchange prefetch. assert_equals: expected "from server" but got "sxg-subresource-script.js should not be prefetched"
Harness: the test ran to completion.
